In this changeset, I am adding a test_sauropod module.
The test runs a fake Sauropod server and runs test_server
against it. This makes sure the sauropod backend is tested
even if there's no Node.js running.

The test also exercise the memcache layer.

Tests should pass if memcache is OFF or ON. OFF ignores
the cache and go through.