Now testing the number of hits to make sure Memcache really caches.
If Memcache is running the tests will use it and make sure it caches
Sauropod GETs on the meta document, and saves hits.

If Memcache is not running, the tests makes sure the backend still
works with no cache.

For the SQL backend, Memcache is mandatory for the tokens, if not
running a verify will generate a 503 and another call a 401.